requirements.txt
alembic==1.4.3
  - Mako [required: Any, installed: 1.1.3]
    - MarkupSafe [required: >=0.9.2, installed: 1.1.1]
  - python-dateutil [required: Any, installed: 2.8.1]
    - six [required: >=1.5, installed: 1.15.0]
  - python-editor [required: >=0.3, installed: 1.0.4]
  - SQLAlchemy [required: >=1.1.0, installed: 1.3.22]
flake8==3.8.4
  - mccabe [required: >=0.6.0,<0.7.0, installed: 0.6.1]
  - pycodestyle [required: >=2.6.0a1,<2.7.0, installed: 2.6.0]
  - pyflakes [required: >=2.2.0,<2.3.0, installed: 2.2.0]
Flask==1.1.2
  - click [required: >=5.1, installed: 7.1.2]
  - itsdangerous [required: >=0.24, installed: 1.1.0]
  - Jinja2 [required: >=2.10.1, installed: 2.11.2]
    - MarkupSafe [required: >=0.23, installed: 1.1.1]
  - Werkzeug [required: >=0.15, installed: 1.0.1]
pipdeptree==2.0.0
  - pip [required: >=6.0.0, installed: 20.2.3]
psycopg2==2.8.6
python-decouple==3.4
setuptools==49.2.1
